+++ src/mesh/cs_mesh.c
@@ -3966,8 +3966,8 @@ cs_mesh_dump(const cs_mesh_t  *mesh)
   if (mesh->n_domains > 1) {
     bft_printf("\nGlobal vertex numbering:\n");
     for (i = 0; i < mesh->n_vertices; i++)
-      bft_printf("   <%7d >  %10u\n",
-                 i+1, mesh->global_vtx_num[i]);
+      bft_printf("   <%7d >  %10llu\n",
+                 i+1, (unsigned long long)(mesh->global_vtx_num[i]));
   }
 
   bft_printf("\n\n        ---------------------------"
@@ -3991,8 +3991,8 @@ cs_mesh_dump(const cs_mesh_t  *mesh)
 
     bft_printf("\nInternal faces global numbering:\n");
     for (i = 0; i < mesh->n_i_faces; i++)
-      bft_printf("   < %7d >  %12d",
-                 i+1, mesh->global_i_face_num[i]);
+      bft_printf("   < %7d >  %12llu",
+                 i+1, (unsigned long long)(mesh->global_i_face_num[i]));
     bft_printf("\n");
 
   }
@@ -4021,7 +4021,8 @@ cs_mesh_dump(const cs_mesh_t  *mesh)
 
     bft_printf("\nCell global numbering:\n");
     for (i = 0; i < mesh->n_cells; i++)
-      bft_printf("   < %7d >  %12d", i+1, mesh->global_cell_num[i]);
+      bft_printf("   < %7d >  %12llu", i+1,
+                 (unsigned long long)(mesh->global_cell_num[i]));
     bft_printf("\n");
 
   }
